refactor: Create a transactions submodule
