+++ editor/server/editor/services/transaction/base.py
@@ -1,0 +1,15 @@
+from abc import ABC, abstractmethod
+
+from editor.models import Program
+
+
+class TransactionBase(ABC):
+    """A transaction that can be accepted from the client to edit a program"""
+
+    @abstractmethod
+    def do(self, original: Program) -> Program:
+        """Apply an operation to the given program and return the resulting program"""
+
+    @abstractmethod
+    def undo(self, original: Program) -> Program:
+        """Undoes the operation performed in `do`"""
